bug fix
bug fix for REG_ID_BAT

The correct power-on sequence for obtaining the battery percentage is to first power on the device with the battery inserted
Then connect the USB Type-C cable to read the battery percentage.
The battery percentage also can be retrieved via I2C communication within the UF2 program.